clean blockchain component
