+++ blockchain/block.go
@@ -6,13 +6,11 @@ import (
 
 	"bytes"
 
-	"github.com/it-chain/it-chain-Engine/core/eventstore"
 	"github.com/it-chain/midgard"
 	ygg "github.com/it-chain/yggdrasill/common"
 
 	"errors"
 	"fmt"
-	"log"
 )
 
 type Block = ygg.Block
@@ -149,34 +147,6 @@ func (block *DefaultBlock) IsPrev(serializedPrevBlock []byte) bool {
 	return bytes.Compare(prevBlock.GetSeal(), block.GetPrevSeal()) == 0
 }
 
-// interface of api gateway query api
-type BlockQueryApi interface {
-	GetBlockByHeight(blockHeight uint64) (Block, error)
-	GetBlockBySeal(seal []byte) (Block, error)
-	GetBlockByTxID(txid string) (Block, error)
-	GetLastBlock() (Block, error)
-}
-
-type Action interface {
-	DoAction(block Block) error
-}
-
-// TODO: Write test case
-func CreateSaveOrSyncAction(checkResult int64) Action {
-	if checkResult > 0 {
-		return NewSyncAction()
-	} else if checkResult == 0 {
-		return NewSaveAction()
-	} else {
-		return NewDefaultAction()
-	}
-}
-
-type SyncAction struct{}
-
-func NewSyncAction() *SyncAction {
-	return &SyncAction{}
-}
 func (block *DefaultBlock) On(event midgard.Event) error {
 
 	switch v := event.(type) {
@@ -203,53 +173,86 @@ func (block *DefaultBlock) On(event midgard.Event) error {
 	return nil
 }
 
-func NewEmptyBlock(prevSeal []byte, height uint64, creator []byte) *DefaultBlock {
-	block := &DefaultBlock{}
-	return block
-}
+//ToDo: 삭제 제안 - junk_sound
+////interface of api gateway query api
+//type BlockQueryApi interface {
+//	GetBlockByHeight(blockHeight uint64) (Block, error)
+//	GetBlockBySeal(seal []byte) (Block, error)
+//	GetBlockByTxID(txid string) (Block, error)
+//	GetLastBlock() (Block, error)
+//}
 
-func (syncAction *SyncAction) DoAction(block Block) error {
-	// TODO: Start synchronize
-	return nil
-}
+// ToDo: 미완성 부분에 대한 주석처리
+//type Action interface {
+//	DoAction(block Block) error
+//}
+//
+//// TODO: Write test case
+//func CreateSaveOrSyncAction(checkResult int64) Action {
+//	if checkResult > 0 {
+//		return NewSyncAction()
+//	} else if checkResult == 0 {
+//		return NewSaveAction()
+//	} else {
+//		return NewDefaultAction()
+//	}
+//}
+//
+//type SyncAction struct{}
+//
+//func NewSyncAction() *SyncAction {
+//	return &SyncAction{}
+//}
 
-type SaveAction struct {
-	blockPool BlockPool
-}
+//ToDo: 삭제 제안 - junk_sound
+//func NewEmptyBlock(prevSeal []byte, height uint64, creator []byte) *DefaultBlock {
+//	block := &DefaultBlock{}
+//	return block
+//}
 
-func NewSaveAction() *SaveAction {
-	return &SaveAction{}
-}
-
-// TODO: Write test case
-func (saveAction *SaveAction) DoAction(block Block) error {
-	event, err := createBlockCommittedEvent(block)
-	if err != nil {
-		return err
-	}
-	blockId := string(block.GetSeal())
-	eventstore.Save(blockId, event)
-	return nil
-}
-
-func createBlockCommittedEvent(block Block) (BlockCommittedEvent, error) {
-	seal := string(block.GetSeal())
-	return BlockCommittedEvent{
-		EventModel: midgard.EventModel{
-			ID: seal,
-		},
-		Seal: seal,
-	}, nil
-}
-
-type DefaultAction struct{}
-
-func NewDefaultAction() *DefaultAction {
-	return &DefaultAction{}
-}
-
-// TODO: Write test case
-func (defaultAction *DefaultAction) DoAction(block Block) error {
-	log.Printf("got shorter height block [%v]", block.GetHeight())
-	return nil
-}
+// ToDo: 미완성 부분에 대한 주석처리
+//func (syncAction *SyncAction) DoAction(block Block) error {
+//	// TODO: Start synchronize
+//	return nil
+//}
+//
+//type SaveAction struct {
+//	blockPool BlockPool
+//}
+//
+//func NewSaveAction() *SaveAction {
+//	return &SaveAction{}
+//}
+//
+//// TODO: Write test case
+//func (saveAction *SaveAction) DoAction(block Block) error {
+//	event, err := createBlockCommittedEvent(block)
+//	if err != nil {
+//		return err
+//	}
+//	blockId := string(block.GetSeal())
+//	eventstore.Save(blockId, event)
+//	return nil
+//}
+//
+//func createBlockCommittedEvent(block Block) (BlockCommittedEvent, error) {
+//	seal := string(block.GetSeal())
+//	return BlockCommittedEvent{
+//		EventModel: midgard.EventModel{
+//			ID: seal,
+//		},
+//		Seal: seal,
+//	}, nil
+//}
+//
+//type DefaultAction struct{}
+//
+//func NewDefaultAction() *DefaultAction {
+//	return &DefaultAction{}
+//}
+//
+//// TODO: Write test case
+//func (defaultAction *DefaultAction) DoAction(block Block) error {
+//	log.Printf("got shorter height block [%v]", block.GetHeight())
+//	return nil
+//}
