+++ blockchain/command_service.go
@@ -1,5 +1,6 @@
 package blockchain
 
-type CommandService interface {
-	SendBlockValidateCommand(block Block) error
-}
+// ToDo: 삭제 제안 - junk_sound
+//type CommandService interface {
+//	SendBlockValidateCommand(block Block) error
+//}
