+++ blockchain/event.go
@@ -14,21 +14,22 @@ type SyncDoneEvent struct {
 	midgard.EventModel
 }
 
-type BlockAddToPoolEvent struct {
-	midgard.EventModel
-	Seal      []byte
-	PrevSeal  []byte
-	Height    uint64
-	TxList    []byte
-	TxSeal    [][]byte
-	Timestamp time.Time
-	Creator   []byte
-}
-
-type BlockRemoveFromPoolEvent struct {
-	midgard.EventModel
-	Height uint64
-}
+// ToDo: 삭제 제안 - junk_sound
+//type BlockAddToPoolEvent struct {
+//	midgard.EventModel
+//	Seal      []byte
+//	PrevSeal  []byte
+//	Height    uint64
+//	TxList    []byte
+//	TxSeal    [][]byte
+//	Timestamp time.Time
+//	Creator   []byte
+//}
+//
+//type BlockRemoveFromPoolEvent struct {
+//	midgard.EventModel
+//	Height uint64
+//}
 
 // event when block is saved to event store
 type BlockCommittedEvent struct {
