+++ blockchain/grpc_command_service.go
@@ -1,7 +1,6 @@
 package blockchain
 
-type GrpcCommandService interface {
-	RequestBlock(peerId PeerId, height uint64) error
-	ResponseBlock(peerId PeerId, block Block) error
-	SyncCheckResponse(block Block) error
-}
+// ToDo: 삭제 제안 - junk_sound
+//type GrpcCommandService interface {
+//	SyncCheckResponse(block Block) error
+//}
